// src/types/companyContact.types.ts
export interface ICompanyContact {
	id: number;
	company_name: string;
	phone: string;
	b_company_name: string;
	b_phone: string;
	b_address: string;
	email: string;
	website: string;
	address: string;
	city: string;
	country: string;
	facebook: string;
	instagram: string;
	youtube: string;
	linkedin: string;
	created_at: string;
	updated_at: string;
	system_date: string;
}